+++ src/http-message/src/Stream/SwooleStream.php
@@ -22,9 +22,14 @@ class SwooleStream implements StreamInterface
     protected $contents;
 
     /**
-     * @var string
+     * @var int
      */
     protected $size;
+
+    /**
+     * @var bool
+     */
+    protected $writable;
 
     /**
      * SwooleStream constructor.
@@ -33,6 +38,7 @@ class SwooleStream implements StreamInterface
     {
         $this->contents = $contents;
         $this->size = strlen($this->contents);
+        $this->writable = true;
     }
 
     /**
@@ -60,7 +66,7 @@ class SwooleStream implements StreamInterface
      */
     public function close()
     {
-        throw new \BadMethodCallException('Not implemented');
+        $this->detach();
     }
 
     /**
@@ -71,7 +77,11 @@ class SwooleStream implements StreamInterface
      */
     public function detach()
     {
-        throw new \BadMethodCallException('Not implemented');
+        $this->contents = '';
+        $this->size = 0;
+        $this->readable = $this->writable = false;
+
+        return null;
     }
 
     /**
@@ -95,7 +105,7 @@ class SwooleStream implements StreamInterface
      */
     public function tell()
     {
-        throw new \BadMethodCallException('Not implemented');
+        throw new \RuntimeException('Cannot determine the position of a SwooleStream');
     }
 
     /**
@@ -105,7 +115,7 @@ class SwooleStream implements StreamInterface
      */
     public function eof()
     {
-        throw new \BadMethodCallException('Not implemented');
+        return $this->getSize() === 0;
     }
 
     /**
@@ -115,7 +125,7 @@ class SwooleStream implements StreamInterface
      */
     public function isSeekable()
     {
-        throw new \BadMethodCallException('Not implemented');
+        return false;
     }
 
     /**
@@ -132,7 +142,7 @@ class SwooleStream implements StreamInterface
      */
     public function seek($offset, $whence = SEEK_SET)
     {
-        throw new \BadMethodCallException('Not implemented');
+        throw new \RuntimeException('Cannot seek a SwooleStream');
     }
 
     /**
@@ -146,7 +156,7 @@ class SwooleStream implements StreamInterface
      */
     public function rewind()
     {
-        throw new \BadMethodCallException('Not implemented');
+        $this->seek(0);
     }
 
     /**
@@ -156,7 +166,7 @@ class SwooleStream implements StreamInterface
      */
     public function isWritable()
     {
-        return false;
+        return $this->writable;
     }
 
     /**
@@ -168,7 +178,16 @@ class SwooleStream implements StreamInterface
      */
     public function write($string)
     {
-        throw new \BadMethodCallException('Not implemented');
+        if (! $this->writable) {
+            throw new \RuntimeException('Cannot write to a non-writable stream');
+        }
+
+        $size = strlen($string);
+
+        $this->contents .= $string;
+        $this->size += $size;
+
+        return $size;
     }
 
     /**
@@ -193,7 +212,17 @@ class SwooleStream implements StreamInterface
      */
     public function read($length)
     {
-        throw new \BadMethodCallException('Not implemented');
+        if ($length >= $this->getSize()) {
+            $result = $this->contents;
+            $this->contents = '';
+            $this->size = 0;
+        } else {
+            $result = substr($this->contents, 0, $length);
+            $this->contents = substr($this->contents, $length);
+            $this->size = $this->getSize() - $length;
+        }
+
+        return $result;
     }
 
     /**